+++ rp2040-hal/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+## Fixed
+
+- Fixed minimum PLL's VCO frequency according to updated datasheet - @ithinuel
+
+### Added
+
+- Support for GPin0 and GPin1 clock sources - @ithinuel
+- Clock frequency approximation function using the frequency counter - @ithinuel
+
 ## [0.9.0]
 
 ### MSRV